ControlAndRbh implemented. sessionState(...) can now wait until all Plants have desired value

+++ constants.go
@@ -1,6 +1,6 @@
 package energontrol
 
-var CtrlValues = map[string]uint32{
+var CtrlValues = map[string]uint64{
 	"Start":              0,
 	"Stop60":             1,
 	"Stop90":             2,
@@ -16,7 +16,7 @@ var RbhValues = map[string]uint64{
 	"PresetDuration": 128,
 }
 
-var sessionStates = map[uint32]string{
+var sessionStates = map[uint16]string{
 	0:   "Session free",
 	1:   "Session reserved",
 	2:   "Parameter input",
@@ -31,25 +31,25 @@ var sessionStates = map[uint32]string{
 }
 
 const (
-	RbhNoAccess                  = uint32(0)       // No Access on Rbh
-	RbhAutoDeicingAllowed        = uint32(1)       // Automatic deicing allowed
-	RbhAutoOffWEA                = uint32(1 << 1)  // Automatic operation of the heater suppressed
-	RbhManualOnWEA               = uint32(1 << 2)  // RBH manually on (inside plant)
-	RbhManualOnSCADA             = uint32(1 << 3)  // RBH manually on (SCADA)
-	RbhAutoDeicingWhenStopped    = uint32(1 << 4)  // Automatic deicing when the system is stopped
-	RbhAutoDeicingInOperation    = uint32(1 << 5)  // Automatic deicing in operation
-	RbhHeatingPreventiveAuto     = uint32(1 << 6)  // Heater preventively turned on (by automation)
-	RbhHeatingWhenStoppedSCADA   = uint32(1 << 7)  // Heater on when the system is stopped (SCADA)
-	RbhHeatingInOperationSCADA   = uint32(1 << 8)  // Heater on during operation (SCADA)
-	RbhNoSupplyPowerAvailable    = uint32(1 << 10) // No supply power available
-	RbhFault                     = uint32(1 << 11) // Heater malfunction
-	RbhDeicingAllowedInOperation = uint32(1 << 12) // Deicing allowed in operation
-	RbhPreventiveHeaterAllowed   = uint32(1 << 13) // Preventive heater allowed
-	RbhInstalled                 = uint32(1 << 15) // Heater installed
-	RbhNotInstalled              = uint32(1 << 16) // Heater not installed
+	RbhNoAccess                  = uint64(0)       // No Access on Rbh
+	RbhAutoDeicingAllowed        = uint64(1)       // Automatic deicing allowed
+	RbhAutoOffWEA                = uint64(1 << 1)  // Automatic operation of the heater suppressed
+	RbhManualOnWEA               = uint64(1 << 2)  // RBH manually on (inside plant)
+	RbhManualOnSCADA             = uint64(1 << 3)  // RBH manually on (SCADA)
+	RbhAutoDeicingWhenStopped    = uint64(1 << 4)  // Automatic deicing when the system is stopped
+	RbhAutoDeicingInOperation    = uint64(1 << 5)  // Automatic deicing in operation
+	RbhHeatingPreventiveAuto     = uint64(1 << 6)  // Heater preventively turned on (by automation)
+	RbhHeatingWhenStoppedSCADA   = uint64(1 << 7)  // Heater on when the system is stopped (SCADA)
+	RbhHeatingInOperationSCADA   = uint64(1 << 8)  // Heater on during operation (SCADA)
+	RbhNoSupplyPowerAvailable    = uint64(1 << 10) // No supply power available
+	RbhFault                     = uint64(1 << 11) // Heater malfunction
+	RbhDeicingAllowedInOperation = uint64(1 << 12) // Deicing allowed in operation
+	RbhPreventiveHeaterAllowed   = uint64(1 << 13) // Preventive heater allowed
+	RbhInstalled                 = uint64(1 << 15) // Heater installed
+	RbhNotInstalled              = uint64(1 << 16) // Heater not installed
 )
 
-var RbhStatus = map[uint32]string{
+var RbhStatus = map[uint64]string{
 	RbhNoAccess:                  "No Access on Rbh",
 	RbhAutoDeicingAllowed:        "Automatic deicing allowed",
 	RbhAutoOffWEA:                "Automatic operation of the heater suppressed",
